% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement_ads.R
\name{agreement_ads}
\alias{agreement_ads}
\title{Agreement's coefficients adjusted}
\usage{
agreement_ads(
  nrt,
  ni,
  na,
  data,
  NC = 4,
  NI = 5,
  data_prior = NULL,
  prior = NULL,
  index = NULL,
  sigma = 0.5,
  sigma.bad = 1,
  uniform = FALSE,
  correct.BT = FALSE
)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.}

\item{NC}{number of chains}

\item{NI}{number of iterations}

\item{data_prior}{dataframe or vector containing the prior}

\item{prior}{weakly informatiove or informative prior}

\item{index}{the inter-rater agreement index to be estimated}

\item{sigma}{Standard Deviation Raters' BMM prior}

\item{sigma.bad}{Standard Deviation Raters' BMM prior based on the difference between posterior predictive and observed data}

\item{uniform}{LOGICAL. If TRUE, the weakly informative prior will follow a uniform distribution.}

\item{correct.BT}{LOGICAL. If TRUE, it corrects the BT estimated value.}
}
\value{
A list containing the following elements:
\itemize{
\item All_BCM - a data frame containing the raw data and the estimated ones, for each rater
\item Results - a data frame containing the following columns:
\enumerate{
\item The Couple of rater compared.
\item The mean BCM of Rater 1.
\item The mean BCM of Rater 2.
\item The BT of Rater 1.
\item The BT of Rater 2.
\item The agreement index before the adjustment.
\item The agreement index after the adjustment.

}
}
}
\description{
This function calculates different agreement's coefficients
between n-couples of raters
}
\examples{
data(toyEx)

agreement_ads(nrt=3,ni=12,na=8,data=toyEx,NC=4,NI=5,prior="WI",index = "SPA",
sigma = .3,sigma.bad=1,uniform=TRUE)
}
\seealso{
\code{toyEx}
}
