% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMM_IP.R
\name{BMM_IP}
\alias{BMM_IP}
\title{Belonging Measure Matrix (BMM) estimation- informative prior}
\usage{
BMM_IP(
  nrt,
  ni,
  na,
  data,
  NC = 4,
  NI = 5,
  data_prior,
  sigma = 0.5,
  correct.BT = FALSE,
  sigma.bad = 1
)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution..}

\item{NC}{Number of chains. Default is equal to four.}

\item{NI}{Number of iteration per chain. Default is equal to five.}

\item{data_prior}{A dataframe containing data representing the informative prior}

\item{sigma}{Standard Deviation Raters' BMM prior}

\item{correct.BT}{LOGICAL. If TRUE, it corrects the BT estimated value.}

\item{sigma.bad}{Standard Deviation Raters' BMM prior based on the difference between posterior predictive and observed data.}
}
\value{
A list containing the following elements:
\itemize{
\item IP - the vector containing the values of the weakly informative prior.
\item BT - the vector containing the Belonging Threshold for each rater.
\item BCM.I - BCM estimated from BMM and BT
\item BadF - the vector containing values reflecting the difference between posterior predictive and observed data
}
}
\description{
This function estimates the Belonging Measure Matrix, starting from
the Belonging Threshold in case of an informative prior
}
\examples{
data(expertBMM)

## Estimate BTs for three rater with a weakly informative prior
expertBMM<-expertBMM[,2:9]
BMM_IP(3,12,8,data=toyEx,data_prior=expertBMM,sigma=.5,correct.BT=FALSE,sigma.bad=1)
}
\seealso{
\code{toyEx}
}
