% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMM_WIP.R
\name{BMM_WIP}
\alias{BMM_WIP}
\title{Belonging Measure Matrix (BMM) estimation-weakly informative prior}
\usage{
BMM_WIP(
  nrt,
  ni,
  na,
  data,
  NC = 4,
  NI = 5,
  sigma = 0.5,
  sigma.bad = 1,
  uniform = FALSE
)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.}

\item{NC}{Number of chains. Default is equal to four.}

\item{NI}{Number of iteration per chain. Default is equal to five}

\item{sigma}{Standard Deviation Raters' BMM prior}

\item{sigma.bad}{Standard Deviation Raters' BMM prior based on the difference between posterior predictive and observed data}

\item{uniform}{LOGICAL. If TRUE, the weakly informative prior will follow a uniform distribution.}
}
\value{
A list containing the following elements:
\itemize{
\item WIP - the vector containing the values of the weakly informative prior.
\item BT - the vector containing the Belonging Threshold for each rater.
\item BCM.WI - BCM estimated from BMM and BT
\item BadF - the vector containing values reflecting the difference between posterior predictive and observed data
}
}
\description{
This function estimates the Belonging Measure Matrix, starting from
the Belonging Threshold in case of a weakly informative prior
}
\examples{
data(toyEx)

## Estimate BTs for three rater with a weakly informative prior
BMM_WIP(3,12,8,data=toyEx,sigma=.5,uniform=TRUE,sigma.bad=1)
}
\seealso{
\code{toyEx}
}
