% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_IP.R
\name{BT_IP}
\alias{BT_IP}
\title{Belonging Threshold (BT) estimation- informative prior}
\usage{
BT_IP(
  nrt,
  ni,
  na,
  data,
  NC = 4,
  NI = 5,
  data_prior,
  sigma = 0.5,
  correct.BT = FALSE
)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Data frame containing the original attribution.}

\item{NC}{Number of chains. Default is equal to four.}

\item{NI}{Number of iteration per chain. Default is equal to five.}

\item{data_prior}{A data frame containing data representing the informative prior}

\item{sigma}{Standard Deviation Raters' BMM prior}

\item{correct.BT}{LOGICAL. If TRUE, it corrects the BT estimated value.}
}
\value{
A list containing the following elements:
\itemize{
\item IP - the vector containing the values of the informative prior.
\item BT - the vector containing the Belonging Threshold for each rater.
\item BadF - the vector containing values reflecting the difference between posterior predictive and observed data
}
}
\description{
This function estimates the Belonging Threshold for each rater
in case of an informative prior
}
\examples{
data(expertBMM)

## Estimate BTs for three rater with a weakly informative prior
expertBMM<-expertBMM[,2:9]
BT_IP(3,12,8,data=toyEx,data_prior=expertBMM,sigma=.5,correct.BT=FALSE)
}
\seealso{
\code{toyEx}
}
