% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_WIP.R
\name{BT_WIP}
\alias{BT_WIP}
\title{Belonging Threshold (BT) estimation-weakly informative prior}
\usage{
BT_WIP(nrt, ni, na, data, NC = 4, NI = 5, sigma = 0.5, uniform = FALSE)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Data frame containing the original attribution..}

\item{NC}{Number of chains. Default is equal to four.}

\item{NI}{Number of iteration per chain. Default is equal to five.}

\item{sigma}{Standard Deviation Raters' BMM prior}

\item{uniform}{LOGICAL. If TRUE, the weakly informative prior will follow a uniform distribution.}
}
\value{
A list containing the following elements:
\itemize{
\item WIP - the vector containing the values of the weakly informative prior.
\item BT - the vector containing the Belonging Threshold for each rater.
\item BadF - the vector containing values reflecting the difference between posterior predictive and observed data
}
}
\description{
This function estimates the Belonging Threshold for each rater
in case of a weakly informative prior
}
\examples{
data(toyEx)

## Estimate BTs for three rater with a weakly informative prior
BT_WIP(3,12,8,data=toyEx,sigma=.5,uniform=TRUE)
}
\seealso{
\code{toyEx}
}
