% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_BCM_list.R
\name{create_BCM_list}
\alias{create_BCM_list}
\title{Create the set of Boolean Classification Matrices (BCM)}
\usage{
create_BCM_list(nrt, ni, na, data)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.
It can be a \code{as_BCM} object.}
}
\value{
A list containing the BCM of each rater.
}
\description{
Given a set of raters (nrt), items (ni) of a scale and their
investigated symptoms/attributes (na), create_BCMM_list() returns
a data list containg the Boolean Classification Matrix (BCM) of each
rater.
}
\examples{
data(toyEx)

create_BCM_list(3,12,8,data=toyEx)
}
\seealso{
\code{as_BCM}, \code{toyEx}
}
