% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expertBMM.R
\docType{data}
\name{expertBMM}
\alias{expertBMM}
\title{Example of Belonging Measure Matrix made by an expert.}
\format{
A data frame with 12 rows and 9 variables:
\describe{
\item{Item}{The items' number of the depression scale. }
\item{A1}{Symptom/attribute investigated by each item.}
\item{A2}{Symptom/attribute investigated by each item.}
\item{A3}{Symptom/attribute investigated by each item.}
\item{A4}{Symptom/attribute investigated by each item.}
\item{A5}{Symptom/attribute investigated by each item.}
\item{A6}{Symptom/attribute investigated by each item.}
\item{A7}{Symptom/attribute investigated by each item.}
\item{A8}{Symptom/attribute investigated by each item.}
}
}
\usage{
data(expertBMM)
}
\description{
A dataset containing a Belonging Measure Matrix
made by an expert of depression. The matrix was performed on a scale
containing twelve items evaluating eight symptoms of depression.
}
\keyword{datasets}
