% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rater_adj.R
\name{multi_rater_adj}
\alias{multi_rater_adj}
\title{Agreement's coefficients adjusted}
\usage{
multi_rater_adj(
  nrt,
  ni,
  na,
  data,
  NC = 4,
  NI = 5,
  sigma = 0.5,
  data_prior = NULL,
  prior,
  index,
  model = NULL,
  s.levels = NULL,
  type = NULL,
  unit = NULL,
  exact = NULL,
  detail = NULL,
  correct = NULL,
  sigma.bad = 1,
  uniform = FALSE,
  correct.BT = FALSE
)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Data frame containing the original attribution.}

\item{NC}{number of chains}

\item{NI}{number of iterations}

\item{sigma}{Standard Deviation Raters' BMM prior}

\item{data_prior}{data frame or vector containing the prior}

\item{prior}{weakly informatiove or informative prior}

\item{index}{the inter-rater agreement index to be estimated}

\item{model}{model}

\item{s.levels}{s_level}

\item{type}{type}

\item{unit}{unit}

\item{exact}{logical}

\item{detail}{details}

\item{correct}{logical}

\item{sigma.bad}{Standard Deviation Raters' BMM prior based on the difference between posterior predictive and observed data}

\item{uniform}{LOGICAL. If TRUE, the weakly informative prior will follow a uniform distribution.}

\item{correct.BT}{LOGICAL. If TRUE, it corrects the BT estimated value.}
}
\value{
A list containing the following elements:
\itemize{
\item All_BCM - a data frame containing the raw data and the estimated ones, for each rater
\item Results - a data frame containing the following columns:
\enumerate{
\item The mean BCM of each Rater.
\item The BT of Rater CM of each Rater.
\item The agreement index before the adjustment.
\item The agreement index after the adjustment.

}
}
}
\description{
This function calculates different agreement's coefficients
between n-couples of raters
}
\examples{
data(toyEx)

## Compute SPA only between rater 1 and 3
multi_rater_adj(3,12,8,data=toyEx,prior="WI",index="kappa.light",
sigma=.5,sigma.bad=1,uniform=TRUE)
}
\seealso{
\code{toyEx}
}
