% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeve-package.R
\docType{package}
\name{sleeve-package}
\alias{sleeve-package}
\alias{sleeve}
\title{The 'sleeve' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
