% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyEx.R
\docType{data}
\name{toyEx}
\alias{toyEx}
\title{Example of item-symptoms attribution matrix.}
\format{
A data frame with 36 rows and 10 variables:
\describe{
\item{Item}{The items' number of the depression scale. }
\item{A1}{Symptom/attribute investigated by each item.}
\item{A2}{Symptom/attribute investigated by each item.}
\item{A3}{Symptom/attribute investigated by each item.}
\item{A4}{Symptom/attribute investigated by each item.}
\item{A5}{Symptom/attribute investigated by each item.}
\item{A6}{Symptom/attribute investigated by each item.}
\item{A7}{Symptom/attribute investigated by each item.}
\item{A8}{Symptom/attribute investigated by each item.}
\item{Subj}{The rater who made the attribution, for each item.}
}
}
\usage{
data(toyEx)
}
\description{
A dataset containing a matrix depicting the relations between a set of
twelve item investigating eight symptoms of a depression scale,
accordingly the opinion of three independent raters.
}
\keyword{datasets}
